+++ SplunkRumWorkspace/SauceLabsTestApp/SauceLabsTestApp/UITests.swift
@@ -1,0 +1,43 @@
+//
+/*
+Copyright 2023 Splunk Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+	
+
+import Foundation
+
+class ViewControllerShowTest: TestCase {
+    init() {
+        super.init(name: "showVC")
+    }
+    
+    override func verify(_ span: TestZipkinSpan) {
+        if span.name != "ShowVC" {
+            return
+        }
+        
+        if span.tags["component"] != "ui" {
+            return self.fail()
+        }
+        
+        if span.tags["last.screen.name"] == nil {
+            return self.fail()
+        }
+        
+        if span.tags["object.type"] == nil {
+            return self.fail()
+        }
+    }
+}
